Drop `V2`: Overview page
