+++ packages/web/app/src/components/project/settings/external-composition.tsx
@@ -4,8 +4,8 @@ import { useMutation, useQuery } from 'urql';
 import * as Yup from 'yup';
 import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
+import { ProductUpdatesLink } from '@/components/ui/docs-note';
 import { DocsNote, Input, Spinner, Switch, Tooltip } from '@/components/v2';
-import { ProductUpdatesLink } from '@/components/v2/docs-note';
 import { FragmentType, graphql, useFragment } from '@/gql';
 import { useNotifications } from '@/lib/hooks';
 import { CheckIcon, Cross2Icon, UpdateIcon } from '@radix-ui/react-icons';
