+++ packages/web/app/src/components/project/settings/native-composition.tsx
@@ -10,8 +10,8 @@ import {
   CardHeader,
   CardTitle,
 } from '@/components/ui/card';
+import { ProductUpdatesLink } from '@/components/ui/docs-note';
 import { useToast } from '@/components/ui/use-toast';
-import { ProductUpdatesLink } from '@/components/v2/docs-note';
 import { FragmentType, graphql, useFragment } from '@/gql';
 import { NativeFederationCompatibilityStatus } from '@/gql/graphql';
