+++ packages/web/app/src/components/ui/card.tsx
@@ -5,7 +5,10 @@ const Card = React.forwardRef<HTMLDivElement, React.HTMLAttributes<HTMLDivElemen
   ({ className, ...props }, ref) => (
     <div
       ref={ref}
-      className={cn('text-card-foreground rounded-lg border bg-gray-900/50 shadow-sm', className)}
+      className={cn(
+        'text-card-foreground rounded-lg border bg-gray-900/50 p-5 shadow-sm',
+        className,
+      )}
       {...props}
     />
   ),
