+++ packages/web/app/src/components/ui/docs-note.tsx
