+++ packages/web/app/src/components/ui/time-ago.tsx
@@ -1,18 +1,39 @@
-import { differenceInWeeks, format, formatDistanceToNow } from 'date-fns';
+import { ReactElement, useMemo } from 'react';
+import clsx from 'clsx';
+import { format } from 'date-fns';
+import { TimeAgo as ReactTimeAgo } from '@n1ru4l/react-time-ago';
 
-export function TimeAgo(props: { date: string; className?: string }) {
-  const date = new Date(props.date);
-  const weeks = differenceInWeeks(new Date(), date);
-  const text =
-    weeks <= 1
-      ? formatDistanceToNow(date, {
-          addSuffix: true,
-        })
-      : format(date, 'PP p');
+export const TimeAgo = ({
+  date,
+  className,
+}: {
+  date?: string;
+  className?: string;
+}): ReactElement | null => {
+  const { dateObj, formattedDate } = useMemo(() => {
+    if (!date) {
+      return {};
+    }
+    const dateObj = new Date(date);
+    const formattedDate = format(dateObj, 'y-MM-dd');
+    return { dateObj, formattedDate };
+  }, [date]);
+
+  if (!date || !dateObj) {
+    return null;
+  }
 
   return (
-    <span className={props.className} title={format(date, 'PP p')}>
-      {text}
-    </span>
+    <ReactTimeAgo date={dateObj}>
+      {({ value }) => (
+        <time
+          dateTime={formattedDate}
+          title={formattedDate}
+          className={clsx('cursor-default whitespace-nowrap', className)}
+        >
+          {value}
+        </time>
+      )}
+    </ReactTimeAgo>
   );
-}
+};
