+++ packages/web/app/src/components/v2/index.ts
@@ -1,5 +1,5 @@
 export { Accordion } from '@/components/v2/accordion';
-export { Activities } from '@/components/v2/activities';
+export { Activities } from '@/components/ui/activities';
 export { Autocomplete } from '@/components/v2/autocomplete';
 export { Avatar } from '@/components/v2/avatar';
 export { Badge } from '@/components/v2/badge';
@@ -10,11 +10,11 @@ export { Checkbox } from '@/components/v2/checkbox';
 export { CopyValue } from '@/components/v2/copy-value';
 export { DataWrapper } from '@/components/v2/data-wrapper';
 export { DiffEditor } from '@/components/v2/diff-editor';
-export { EmptyList, noSchema } from '@/components/v2/empty-list';
+export { EmptyList, noSchema } from '@/components/ui/empty-list';
 export { GraphQLBlock } from '@/components/v2/graphql-block';
 export { Heading } from '@/components/v2/heading';
 export { Input } from '@/components/v2/input';
-export { Link } from '@/components/v2/link';
+export { Link } from '@/components/ui/link';
 export { Modal } from '@/components/v2/modal';
 export { ProjectTypes } from '@/components/v2/project-types';
 export { Radio, RadioGroup } from '@/components/v2/radio-group';
@@ -31,7 +31,7 @@ export { Switch } from '@/components/v2/switch';
 export { Table, TBody, THead, TFoot, Th, Td, Tr } from '@/components/v2/table';
 export { Tabs } from '@/components/v2/tabs';
 export { Tag } from '@/components/v2/tag';
-export { TimeAgo } from '@/components/v2/time-ago';
+export { TimeAgo } from '@/components/ui/time-ago';
 export { ToggleGroup, ToggleGroupItem } from '@/components/v2/toggle-group';
 export { Tooltip } from '@/components/v2/tooltip';
-export { DocsNote, DocsLink } from '@/components/v2/docs-note';
+export { DocsNote, DocsLink } from '@/components/ui/docs-note';
