+++ packages/web/app/src/pages/organization-policy.tsx
@@ -4,11 +4,11 @@ import { OrganizationLayout, Page } from '@/components/layouts/organization';
 import { PolicySettings } from '@/components/policy/policy-settings';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
 import { Checkbox } from '@/components/ui/checkbox';
+import { DocsLink, DocsNote } from '@/components/ui/docs-note';
 import { Meta } from '@/components/ui/meta';
 import { Subtitle, Title } from '@/components/ui/page';
 import { QueryError } from '@/components/ui/query-error';
 import { useToast } from '@/components/ui/use-toast';
-import { DocsLink, DocsNote } from '@/components/v2/docs-note';
 import { graphql } from '@/gql';
 import { OrganizationAccessScope, RegistryModel } from '@/gql/graphql';
 import { useOrganizationAccess } from '@/lib/access/organization';
