+++ packages/web/app/src/pages/organization-settings.tsx
@@ -12,11 +12,11 @@ import {
   CardHeader,
   CardTitle,
 } from '@/components/ui/card';
+import { DocsLink } from '@/components/ui/docs-note';
 import { Meta } from '@/components/ui/meta';
 import { Subtitle, Title } from '@/components/ui/page';
 import { QueryError } from '@/components/ui/query-error';
 import { useToast } from '@/components/ui/use-toast';
-import { DocsLink } from '@/components/v2/docs-note';
 import { GitHubIcon, SlackIcon } from '@/components/v2/icon';
 import { Input } from '@/components/v2/input';
 import {
