+++ packages/web/app/src/pages/project-settings.tsx
@@ -16,11 +16,11 @@ import {
   CardHeader,
   CardTitle,
 } from '@/components/ui/card';
+import { DocsLink } from '@/components/ui/docs-note';
 import { Meta } from '@/components/ui/meta';
 import { Subtitle, Title } from '@/components/ui/page';
 import { QueryError } from '@/components/ui/query-error';
 import { useToast } from '@/components/ui/use-toast';
-import { DocsLink } from '@/components/v2/docs-note';
 import { HiveLogo } from '@/components/v2/icon';
 import { Input } from '@/components/v2/input';
 import { DeleteProjectModal } from '@/components/v2/modals';
