+++ packages/web/app/src/pages/target-checks-single.tsx
@@ -10,20 +10,20 @@ import {
   NoGraphChanges,
 } from '@/components/target/history/errors-and-changes';
 import { Button } from '@/components/ui/button';
+import { DocsLink } from '@/components/ui/docs-note';
+import { EmptyList } from '@/components/ui/empty-list';
 import { Meta } from '@/components/ui/meta';
 import { Subtitle, Title } from '@/components/ui/page';
 import { Popover, PopoverArrow, PopoverContent, PopoverTrigger } from '@/components/ui/popover';
 import { ScrollArea } from '@/components/ui/scroll-area';
 import { Tabs, TabsList, TabsTrigger } from '@/components/ui/tabs';
 import { Textarea } from '@/components/ui/textarea';
+import { TimeAgo } from '@/components/ui/time-ago';
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui/tooltip';
 import { DiffEditor } from '@/components/v2/diff-editor';
-import { DocsLink } from '@/components/v2/docs-note';
-import { EmptyList } from '@/components/v2/empty-list';
 import { Heading } from '@/components/v2/heading';
 import { AlertTriangleIcon, DiffIcon } from '@/components/v2/icon';
 import { Spinner } from '@/components/v2/spinner';
-import { TimeAgo } from '@/components/v2/time-ago';
 import { FragmentType, graphql, useFragment } from '@/gql';
 import { CriticalityLevel, ProjectType } from '@/gql/graphql';
 import { cn } from '@/lib/utils';
