+++ packages/web/app/src/pages/target-checks.tsx
@@ -1,16 +1,16 @@
 import { useState } from 'react';
 import { useQuery } from 'urql';
 import { Page, TargetLayout } from '@/components/layouts/target';
+import { DocsLink } from '@/components/ui/docs-note';
+import { EmptyList } from '@/components/ui/empty-list';
 import { Label } from '@/components/ui/label';
 import { Meta } from '@/components/ui/meta';
 import { Subtitle, Title } from '@/components/ui/page';
 import { QueryError } from '@/components/ui/query-error';
 import { Switch } from '@/components/ui/switch';
+import { TimeAgo } from '@/components/ui/time-ago';
 import { Badge } from '@/components/v2/badge';
 import { Button as LegacyButton } from '@/components/v2/button';
-import { DocsLink } from '@/components/v2/docs-note';
-import { EmptyList } from '@/components/v2/empty-list';
-import { TimeAgo } from '@/components/v2/time-ago';
 import { graphql } from '@/gql';
 import { cn } from '@/lib/utils';
 import { ExternalLinkIcon } from '@radix-ui/react-icons';
