+++ packages/web/app/src/pages/target-insights-coordinate.tsx
@@ -16,11 +16,11 @@ import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert';
 import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
 import { DateRangePicker, presetLast7Days } from '@/components/ui/date-range-picker';
+import { EmptyList } from '@/components/ui/empty-list';
+import { Link as LegacyLink } from '@/components/ui/link';
 import { Meta } from '@/components/ui/meta';
 import { Subtitle, Title } from '@/components/ui/page';
 import { QueryError } from '@/components/ui/query-error';
-import { EmptyList } from '@/components/v2/empty-list';
-import { Link as LegacyLink } from '@/components/v2/link';
 import { CHART_PRIMARY_COLOR } from '@/constants';
 import { graphql } from '@/gql';
 import { formatNumber, formatThroughput, toDecimal } from '@/lib/hooks';
