+++ packages/web/app/src/pages/target-insights.tsx
@@ -10,10 +10,10 @@ import { OperationsList } from '@/components/target/insights/List';
 import { OperationsStats } from '@/components/target/insights/Stats';
 import { Button } from '@/components/ui/button';
 import { DateRangePicker, presetLast7Days } from '@/components/ui/date-range-picker';
+import { EmptyList } from '@/components/ui/empty-list';
 import { Meta } from '@/components/ui/meta';
 import { Subtitle, Title } from '@/components/ui/page';
 import { QueryError } from '@/components/ui/query-error';
-import { EmptyList } from '@/components/v2/empty-list';
 import { graphql } from '@/gql';
 import { useDateRangeController } from '@/lib/hooks/use-date-range-controller';
